requirements.txt
click==7.1.2
fastapi==0.61.1
h11==0.9.0
httptools==0.1.1
numpy==1.19.1
opencv-python-headless==4.4.0.42
Pillow==7.2.0
pydantic==1.6.1
PyYAML==5.3.1
starlette==0.13.6
tflite-runtime==2.1.0.post1
uvicorn==0.11.8
uvloop==0.14.0
websockets==8.1
